+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix an autocorrect error for `RSpec/ExpectActual`. ([@bquorning])
 - Add new `RSpec/UndescriptiveLiteralsDescription` cop. ([@ydah])
+- Add new `RSpec/EmptyOutput` cop. ([@bquorning])
 
 ## 2.28.0 (2024-03-30)